refactor: split repo metrics modules
