+++ docs/metrics/repo-metrics.md
@@ -7,10 +7,10 @@ Generated by `aidevops metrics generate`.
 
 | Metric | Value |
 |---|---:|
-| Source files | 5,194 |
-| Lines of code | 1,008,329 |
-| Comment lines | 147,360 |
-| Blank lines | 206,559 |
+| Source files | 5,204 |
+| Lines of code | 1,008,411 |
+| Comment lines | 147,390 |
+| Blank lines | 206,584 |
 | Languages | 15 |
 | Direct dependencies | 116 |
 | Locked/total dependencies | 116 |
@@ -22,7 +22,7 @@ Generated by `aidevops metrics generate`.
 | Shell | 550,502 | 1,675 | 54.6% |
 | Markdown | 239,682 | 2,891 | 23.8% |
 | HTML | 126,974 | 54 | 12.6% |
-| Python | 24,727 | 138 | 2.5% |
+| Python | 24,809 | 148 | 2.5% |
 | JavaScript | 20,884 | 129 | 2.1% |
 | JSON | 14,560 | 110 | 1.4% |
 | TypeScript | 10,704 | 77 | 1.1% |
